+++ library/SolverUtils/Filters/FilterFieldConvert.h
@@ -76,15 +76,15 @@ public:
 protected:
     SOLVER_UTILS_EXPORT virtual void v_Initialise(
         const Array<OneD, const MultiRegions::ExpListSharedPtr> &pFields,
-        const NekDouble &time);
+        const NekDouble &time) override;
     SOLVER_UTILS_EXPORT virtual void v_FillVariablesName(
         const Array<OneD, const MultiRegions::ExpListSharedPtr> &pFields);
     SOLVER_UTILS_EXPORT virtual void v_Update(
         const Array<OneD, const MultiRegions::ExpListSharedPtr> &pFields,
-        const NekDouble &time);
+        const NekDouble &time) override;
     SOLVER_UTILS_EXPORT virtual void v_Finalise(
         const Array<OneD, const MultiRegions::ExpListSharedPtr> &pFields,
-        const NekDouble &time);
+        const NekDouble &time) override;
     SOLVER_UTILS_EXPORT virtual void v_ProcessSample(
         const Array<OneD, const MultiRegions::ExpListSharedPtr> &pFields,
         std::vector<Array<OneD, NekDouble>> &fieldcoeffs,
@@ -109,7 +109,7 @@ protected:
         const Array<OneD, const MultiRegions::ExpListSharedPtr> &pFields,
         int dump = -1);
 
-    SOLVER_UTILS_EXPORT virtual bool v_IsTimeDependent();
+    SOLVER_UTILS_EXPORT virtual bool v_IsTimeDependent() override;
 
     void CreateModules(std::vector<std::string> &modcmds);