+++ library/SolverUtils/Filters/FilterThresholdMax.h
@@ -72,17 +72,17 @@ protected:
     /// Initialises the filter.
     virtual void v_Initialise(
         const Array<OneD, const MultiRegions::ExpListSharedPtr> &pFields,
-        const NekDouble &time);
+        const NekDouble &time) override;
     /// For each point in domain test if solution is above threshold.
     virtual void v_Update(
         const Array<OneD, const MultiRegions::ExpListSharedPtr> &pFields,
-        const NekDouble &time);
+        const NekDouble &time) override;
     /// Finalise the filter and write out data.
     virtual void v_Finalise(
         const Array<OneD, const MultiRegions::ExpListSharedPtr> &pFields,
-        const NekDouble &time);
+        const NekDouble &time) override;
     /// Indicate that this filter is time dependent.
-    virtual bool v_IsTimeDependent();
+    virtual bool v_IsTimeDependent() override;
 
 private:
     /// Storage for recording when each point in domain rises above threshold.