+++ library/SolverUtils/UnsteadySystem.h
@@ -151,23 +151,20 @@ protected:
         const SpatialDomains::MeshGraphSharedPtr &pGraph);
 
     /// Init object for UnsteadySystem class.
-    SOLVER_UTILS_EXPORT virtual void v_InitObject(bool DeclareField = true);
+    SOLVER_UTILS_EXPORT virtual void v_InitObject(
+        bool DeclareField = true) override;
 
     /// Get the maximum timestep estimator for cfl control.
     SOLVER_UTILS_EXPORT NekDouble MaxTimeStepEstimator();
 
     /// Solves an unsteady problem.
-    SOLVER_UTILS_EXPORT virtual void v_DoSolve();
+    SOLVER_UTILS_EXPORT virtual void v_DoSolve() override;
 
     /// Sets up initial conditions.
-    SOLVER_UTILS_EXPORT virtual void v_DoInitialise();
+    SOLVER_UTILS_EXPORT virtual void v_DoInitialise() override;
 
     /// Print a summary of time stepping parameters.
-    SOLVER_UTILS_EXPORT virtual void v_GenerateSummary(SummaryList &s);
-
-    /// Print the solution at each solution point in a txt file
-    SOLVER_UTILS_EXPORT virtual void v_AppendOutput1D(
-        Array<OneD, Array<OneD, NekDouble>> &solution1D);
+    SOLVER_UTILS_EXPORT virtual void v_GenerateSummary(SummaryList &s) override;
 
     SOLVER_UTILS_EXPORT virtual NekDouble v_GetTimeStep(
         const Array<OneD, const Array<OneD, NekDouble>> &inarray);
@@ -196,6 +193,10 @@ protected:
 
 private:
     void InitializeSteadyState();
+
+    /// Print the solution at each solution point in a txt file
+    SOLVER_UTILS_EXPORT void AppendOutput1D(
+        Array<OneD, Array<OneD, NekDouble>> &solution1D);
 
     bool CheckSteadyState(int step);
     bool CheckSteadyState(int step, NekDouble totCPUTime);