+++ CHANGELOG.md
@@ -29,6 +29,7 @@ v5.3.0
 - Add command line option and environment variable to disable backup field files (!1154)
 - Add override keyword to virtual functions in SpatialDomains (!1448)
 - Add missing override keyword to virtual functions in Collections (!1453)
+- Add missing override keyword to virtual functions in SolverUtils (!1451)
 
 **Python**
 - Add wrappers for Interpreter and Equation classes (!1329)